# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CeanX
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://backend.ceanx.lubomedia.de'
paths:
...
  /api/cost_estimations/latest:
    get:
      summary: 'Get latest Cost Estimations.'
      operationId: getLatestCostEstimations
      description: 'This endpoint allows you to retrieve the latest 5 Cost Estimation.'
      parameters:
        -
          in: header
          name: X-CSRF-TOKEN
          description: ''
          example: '{TOKEN}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - 'Cost Estimations'
...
  '/api/cost_estimations/{id}':
    get:
      summary: 'Get specific Cost Estimation.'
      operationId: getSpecificCostEstimation
      description: 'This endpoint allows you to retrieve a specific Cost Estimation.'
      parameters:
        -
          in: header
          name: X-CSRF-TOKEN
          description: ''
          example: '{TOKEN}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - 'Cost Estimations'
    delete:
      summary: 'Delete Cost Estimation.'
      operationId: deleteCostEstimation
      description: 'This endpoint allows you to delete a specific Cost Estimation.'
      parameters:
        -
          in: header
          name: X-CSRF-TOKEN
          description: ''
          example: '{TOKEN}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - 'Cost Estimations'
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the cost estimation.'
        example: 1
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: apiKey
      name: X-CSRF-TOKEN
      in: header
      description: ''
security:
  -
    default: []